+++ CHANGELOG.md
@@ -1,4 +1,9 @@
 
+## 1.0.4
+- Fixed [#22](https://github.com/uproid/finch/issues/22) fixed checking isApiEndpoint property in WebRequest class
+- Disabled MongoDB by default in FinchConfig [#25](https://github.com/uproid/finch/issues/25)
+- Developed FinchApp.cleanRoutes() method to clear all registered routes [#15](https://github.com/uproid/finch/issues/15)
+
 ## 1.0.3
 - Improved runing app speed in production mode #18